write to temporary files and move
Make saving of the histories atomic by using write to temporary file
plus atomic move trick.[1]

Note, that this is not super-duper solution to all possible issues
related to file saving [2].  But temporary file + move is a good enough
for this app.

While we're here, print IOException message and stacktrace to stderr.

[1] https://softwareengineering.stackexchange.com/a/153175/157607
[2] https://danluu.com/file-consistency/