autosave histories every ten minutes
To avoid information loss in case the app is closed abnormally, save the
histories automatically every ten minutes.

We maybe do not need to stop() the autoSaveTimer when window is closed,
but let's do it anyway, in case this window ever stops being the main
window of the application.  In that case, we shouldn't run the timer
without the GUI, which allows changing the data being saved.

Possible premature optimization: save only those files, for which the
histories have changed.  It's premature, because the .habit files are
very small, it takes a very small amount of time to save them.  On the
other hand, keeping track of which histories have been edited since the
last save point is unnecessary complexity at this point.