create temp file in the destination directory
When saving using a write+rename trick, creating a temporary files in
the default location might put the temporary file on another filesystem.
In most operating systems, atomic moves from one filesystem to another
are not supported, thus causing an AtomicMoveNotSupportedException.

So create the temporary file in the destination directory of the state
file instead.  This also means that in the case of failure during
saving, the user will be able to see the temporary file in the same
directory as main files.  Since the app opens all files with .habit
extension, distinguish the temporary file with '.tmp' suffix.