AboutDialog: check if Desktop.Action.BROWSE is supported
Class Desktop throws UnsupportedOperationException from its methods, if
a particular operation is not supported on the platform, including its
browse() method.

So chack if Desktop.Action.BROWSE is supported before calling browse()
and catch UnsupportedOperationException just in case.  As a fallback for
cases where browser can not be invoked directly, show the user a
JTextField, from which the user can copy the URL.