EditHabitsDialog.HabitListLayout: calculate width of up & down buttons from height of rows
Fix a bug in method `getSizeByGetter`.  When `restWidth` is calculated,
the switch for enum values `MOVE_UP` and `MOVE_DOWN` incorrectly uses
height of the whole dialog to calculate the width.  Refactor how height
is calculated and use the actual `rowHeight` when calculating the width
of the buttons which we are forcing to be squares.