ChooseHabitNameDialog: pass "Window parent" explicitly
The way method "SwingUtilities.getWindowAncestor" works is that it
doesn't work as I expected for parameters of class "Window". It calls
"getParent()" on any parameters, which means that it skips the
opportunity to just return its input parameter.

to avoid this quirk of the SwingUtilities API and just pass Window into
the class ChooseHabitNameDialog.