Gradle: switch to "fat jar" distribution
Gradle plugin 'application' is nice, but it is mostly targeted towards
server deployments of JVM applications.  For GUI application with no
installer, forcing users to find `bin/resoday.{bat,sh}` script is
suboptimal.

So build Resoday, the application, using the "fat jar" approach.  That
is, all compiled classes and classes of all dependencies are gathered
into a single jar file.  To distinguish it from regular `resoday.jar`,
add `-release` suffix to the fat jar.  Drop the integration with the
Gradle plugin 'application', which is no longer needed.

Note, that it is possible that Resoday will have to revert to using
'application' plugin again in the future.