AboutDialog: show URL open fallback in separate JOptionPane
In the next commit, URL stuff is going to be extracted into a separate
class from AboutDialog.  The way JTextField is used as a fallback would
not be a good fit for API of that class, so replace it with something
more agnostic of the layout of the component receiving the clicks.