AboutDialog: add tab "License"
Add tab which explains that Resoday is using GPLv3 license and includes
the whole text of GPLv3 to the AboutDialog.  I couldn't get the resource
file 'LICENSE.txt' to not duplicate the 'LICENSE.txt' in the root of the
repository, but it doesn't matter too much -- the text of the license is
not likely to change.

The stupid experiment in tabs.addChangeListener is for showing more of
the GPLv3 text when clicking on the "License" tab.  It's probably not
worth it.