AboutDialog: specify UTF-8 as charset in content type
HTML parsing in Swing depends on default charset [1].  When a default
charset is not UTF-8, it causes rendering issues in Unicode symbols
copyright '©' and ndash '–'.

So specify the charset explicitly in the content type for JTextPanes
with HTML content.

An alternative solution would be to use '-Dfile.encoding=UTF-8' on the
command line.  This is hard to do with single-jar distribution model.
We could just revert the fat jar commit and tweak the Gradle plugin
'application' to add '-Dfile.encoding=UTF-8' to the command line of the
generated launch scripts (shell and .bat).

Reported-by: Soviet666
Fixes https://github.com/rybak/resoday/issues/1

[1]: the return value of System.getProperty("file.encoding");