MainGui: "Edit habits" -> "Reorder habits"
At some point in the future, "Rename habit" will become "Edit habit"
(with description, settings, etc). "Edit habits" already is a bit weird.
"Reorder" is be better.  So change the text of the menu item in the menu
of MainGui to "Reorder habits".  Rename the class and corresponding
methods and variables too.  Also, update the plain wrong Javadoc at the
top of ReorderHabitsDialog.