Gradle: calculate version string from tags
Having a version in the releaseJar is useful for all kinds of reasons.
Emulate git.git's shell script "GIT-VERSION-GEN" in Groovy/Gradle and
generate the version string from annotated tags.

In Java code, programmatically extract the version and supply it in
DebugDialog for now.  In future commit, it will be added to AboutDialog.

Note, that the new classpath in IDEA run config "Resoday testing from
releaseJar" will take any and all jar files in the directory
"build/distributions".  That's why running Gradle task "clean" is needed
in that run config before creating the jar file.