README: instruct users to download from releases
Now that Resoday 1.0 has been released and `resoday-release-v1.0.jar`
was uploaded to GitHub's "Releases" tab, tell users how to get the jar
file in the README.  Drop the sentence about "make a build for
yourself".  Reorder sections so that section "Usage", intended for users
is higher that "Building", intended for developers.