CalendarPanel: introduce interface for different layouts
In preparation for adding a user setting for changing the layout of the
calendar (buttons, month labels, and other GUI elements), extract the
code for the layout into new interface CalendarPanel with a single
implementation -- ColumnsCalendarPanel.

Note, that highlighting of today's button remains in YearPanel, but
individual implementations of CalendarPanel now need to make all buttons
unhighlighted by default.