remove listeners from YearHistory when removing HistoryPanels
When method `dev.andrybak.resoday.gui.Histories#reorder` juggles
`HistoryPanel`s around, it leaves listeners in `YearHistory`s, which are
added by constructor of `HistoryPanel`.  This leads to a memory leak and
ot a user-facing bug: the user hears the audio from `AudioPlayer` as
many times as they have edited the order of habits.

So remove listeners added to `YearHistory` instance by `HistoryPanel`
when removing the panel from GUI.