WindowPosition: turn into a record class
Because class WindowPosition has only final fields, it can be converted
into a record class.  Do it automatically, via IntelliJ IDEA's
inspection.  Note, that it is possible, because Gson isn't used for its
serialization -- WindowPosition instances are stored and loaded to and
from file via custom `save` and `read` methods.