Gradle: releaseJar: reduce indirection for dependencies
Gradle task releaseJar packages all dependencies into one fat jar by
iterating over them using method collect() and passing `zipTree`s of
every jar [1] to the method from() of class AbstractCopyTask.  On the
code level, however, the method is called with a Closure as a parameter.

Call method from() with the result of method collect() directly rather
than with a Closure that returns the result of method collect().

[1] Technically, the ternary operator isn't needed -- all existing
    dependencies are jar files right now.