MainGui: fix blurry icon on Windows
For some reason, the 64 by 64 pixels icon doesn't get downscaled
correctly on Windows in the top left corner of the JFrame.  Use method
setIconImages() that takes several `Image`s of different resolutions to
address this issue.

TODO: check if setIconImages() addresses the Windows blurriness issue.

An alternative solution would be to set a BaseMultiResolutionImage [1]
that consists of a 32x32 icon and a 64x64 icon as the icon for the
JFrame to fix the icon scaling/rendering issue.  However, there is a JDK
bug in Linux that breaks BaseMultiResolutionImage, see

    https://github.com/rybak/foobar/blob/master/baz/src/main/java/swing/MultiResolutionImageBug.java

On Windows, on the other hand, BaseMultiResolutionImage is rendered
differently to ToolkitImage.  Windows uses an icon of size 56x56 pixels.
When BaseMultiResolutionImage is passed, an image is constructed by
drawing my 64x64 variant as a 48x48 image into a BufferedImage of size
56x56.  See method sun.awt.SunToolkit#getScaledIconImage.

[1] BaseMultiResolutionImage is from Java 9, while setIconImages() is
    from Java 6.