MainGui: set icon for Taskbar
Use java.awt.Taskbar added in Java 9 to set the taskbar icon image. This
mainly affects macOS, as far as I know, which doesn't automatically take
the icon of the JFrame (`window`).

The default Duke icon will be replaced by the Resoday logo in the macOS
dock.  The downside is that macOS doesn't segregate different Java
applications, so all apps running under with the same binary file `java`
are gathered under the one icon.