implement per-habit calendar layouts
Commit 69ed959 (GuiSettings: introduce settings for GUI, 2022-04-09)
introduced a way to change calendar layout for the whole application.

Introduce a similar per-habit setting, which allows more granular
control of how a user wants to navigate the calendar.  The old setting
for the whole application is used as a default fallback for all habits,
for which the user didn't select a custom layout.