MainGui: fix IndexOutOfBoundsException
When habits are reordered, MainGui attempts to find the index of the
tab, which was previously open.  If that tab has been hidden during
reordering, then MainGui opens the first tab (index=0).  However, it is
possible to get an IndexOutOfBoundsException on this setSelectedIndex(0)
call, if all habits were hidden.

Fix this exception by checking the amount of tabs first.