DebugDialog: add screen and window size
A bug report [1] has shown that it would be useful to include the screen
size and window size in the debug information generated by DebugDialog.

The toString of class java.awt.Dimension is too verbose, so implement
our own toString method with just an "x" between the width and height.

[1] "Incorrect display - buttons don't fit in the window vertically"
    https://github.com/rybak/resoday/issues/4