YearPanel: wrap CalendarPanel in a JScrollPane
On resolutions with height less than 1080 the buttons of the
CalendarPanel can get squashed, with no way to access them without a
taller screen.

Wrap the CalendarPanel in a JScrollPane to make the buttons visible when
the window doesn't have room for the calendar buttons.  Set the empty
border to the JScrollPane to override the default border.  This way,
when there is enough room, the GUI is cleaner and doesn't look as if it
wrapped in a JScrollPane.

Closes https://github.com/rybak/resoday/issues/4
       "Incorrect display - buttons don't fit in the window vertically"

Reported-by: GrimPixel